Refactor `buildTree` to use new entity creation utilities
