+++ src/h5web/visualizations/index.test.ts
@@ -9,7 +9,7 @@ import {
   stringType,
   makeStrAttr,
   floatType,
-} from '../providers/mock/data';
+} from '../providers/raw-utils';
 import {
   makeMyDataset,
   makeMyDatatype,
